LayoutTests/imported/w3c/web-platform-tests/css/filter-effects/parsing/color-interpolation-filters-parsing-valid-expected.txt


PASS e.style['color-interpolation-filters'] = "auto" should set the property value
FAIL e.style['color-interpolation-filters'] = "sRGB" should set the property value assert_equals: serialization should be canonical expected "srgb" but got "sRGB"
FAIL e.style['color-interpolation-filters'] = "linearRGB" should set the property value assert_equals: serialization should be canonical expected "linearrgb" but got "linearRGB"
FAIL e.style['color-interpolation-filters'] = "LiNeArRgB" should set the property value assert_equals: serialization should be canonical expected "linearrgb" but got "linearRGB"
